chore(tests): reorganise tests tree
Otherwise, cargo test will try to run each .rs file within the tests folder
which will fail as many of them are intended to be submodules.
