+++ tests/example_tests.rs
@@ -1,0 +1,5 @@
+mod util;
+
+mod examples {
+    mod hello_world;
+}
